fix mistakes

+++ src/Applied/PhoneBook/PhoneBookValidator.java
@@ -5,13 +5,13 @@ import java.util.Set;
 
 public final class PhoneBookValidator {
 
-    public static void checkName(String name, Map<String, Set<Integer>> phoneBook){
+    public static void checkName(String name, Map<String, Set<Long>> phoneBook){
         if(!phoneBook.containsKey(name)){
             throw new IllegalArgumentException("Такого имени нет в телефонной книге");
         }
     }
 
-    public static void checkNumber(String name, Integer number, Map<String, Set<Integer>> phoneBook){
+    public static void checkNumber(String name, Long number, Map<String, Set<Long>> phoneBook){
         if(!phoneBook.get(name).contains(number)){
             throw new IllegalArgumentException("Номер телефона не существует для имени: " + name);
         }
